search_ops_app/requirements.txt
black==22.3.0
db-dtypes==1.0.0
flake8==4.0.1
google-cloud-bigquery==3.0.1
grpcio-status==1.44.0
mkdocs==1.3.0
mkdocs-click==0.5.0
mkdocs-git-revision-date-localized-plugin==1.0.0
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-jupyter==0.20.1
mypy==0.942
pip-chill==1.0.1
plotly==5.7.0
pylint==2.13.7
st-card-component==0.0.10
streamlit-aggrid==0.2.3.post2
streamlit-folium==0.6.5
streamlit==1.9.0
pandas==1.4.2
Jinja2==3.1.2
PyYAML==6.0
python-box==6.0.2
numpy==1.22.3
pytest==7.1.2